replace Puma with Netty
Implement an HTTP Server with Netty, replacing Puma.
This new implementation produces an identical event structure as before
and provides the same keystore mechanisms for TLS so it doesn't break bwc.

To deal with backpressure, this implementation provides two thread pools
of size :threads.
One is responsible for accepting connections, reading http requests
from the established channels and pushing them into an ArrayBlockingQueue.
The other takes requests from the queue, processes them and pushes them
to the logstash queue.