Prevent file descriptor leaks by limiting the max number of connections with un-returned responses.
This plugin currently leaks file descriptors. This patch prevents that by changing the behavior such that when all but one webserver threads are occupied a 429 error will be immediately returned. This means that in the event of a blocked queue we will have one thread available to reject all new attempts to write.

This isn't perfect, we need queue write timeouts to do this well, but it's a decent short term fix.