avoid creation of SSLContext for each connection
Separate the creation of the SslContext and the SslHandler.
Before, for each new connection an SslContext was created and
from it the SslHandler was generated.
However creating an SslContext is very expensive due to all the certificate handling.

This PR moves the creation of SslHandler to a SslHandlerProvider that is instantiated
with a SslContext instance.
This way the SslContext is created once, and from it we get SslHandlers for each new connection.

This largely improves the performance in usecases where ssl is enabled but http keepalive isn't set.