Handle JCE restrictions
While it is been the default for Java since JDK 8u161 released in early 2018,
old versions of Java may not have the JCE unlimited strength jurisdiction policy
installed. This commit handles this case, warning the user that the policy is not
installed, and presenting a reduced set of default ciphers for use.