Fix flaky assertion bound in testThreadSafetyBruteForce
Local compaction depends on completion order and timing. Under certain cases
node count can reach or exceed `sqrt(concurrency)` before the final compaction
is triggered. The test for intermediate state was failing fairly often in CI.
This commit updates the test for intermediate state to simply be less than
`concurrency`. This should satisfy the intent of the test, with the added
security of the rest of the test asserting the finalized state is less than 2
nodes.